feat: implement fractional shares support with u64 quantity types [skip ci]
- Refactor Qty type from u32 to u64 to support fractional shares with basis points precision
- Add complete AccountService implementation with database integration
- Implement core types: Balance, Position, Trade, Reservation with fractional share support
- Add SQL migrations for accounts, positions, trades, reservations tables
- Update all quantity fields throughout the system (whistle, order-gateway, execution-manager, symbol-coordinator, admin-cli)
- Fix type mismatches and arithmetic operations for u64 quantities
- All tests passing (16/16 AccountService, 51/51 Whistle, 25/25 ExecutionManager, etc.)

This enables high-precision fractional share trading with 1/10,000th precision using basis points.

+++ engine/execution-manager/src/lib.rs
@@ -414,7 +414,7 @@ impl ExecutionManager {
                     side: side_str.to_string(),
                     order_type: order_type_str.to_string(),
                     price: submitted.price.map(|p| p as u64), // Convert u32 to u64
-                    quantity: submitted.quantity as u64,      // Convert u32 to u64
+                    quantity: submitted.quantity,
                     order_id: submitted.order_id,
                 }
             }
